[Add] Add netty/xmemcached/jedis/lettuce-core dependency [Update] Update bcprov-jdk15on to 1.66 [Update] Update snmp4j to 2.8.5 [Update] Update jackson to 2.11.3 [Update] Update jakarta.servlet-api to 4.0.4

+++ src/main/java/org/nervousync/utils/CRCUtils.java
@@ -337,7 +337,7 @@ public final class CRCUtils {
 		 */
 		public void appendBinary(byte[] dataBytes) {
 			for (byte dataByte : dataBytes) {
-				int crc = (dataByte < 0) ? ((int)dataByte) + 256 : dataByte;
+				long crc = (dataByte < 0) ? ((int)dataByte) + 256 : dataByte;
 				if (this.crcConfig.getBit() <= 8) {
 					this.crc ^= crc;
 				} else {
